python-utils/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
nltk==3.4.4
pkg-resources==0.0.0
pymongo==3.10.1
python-dotenv==0.10.3
requests==2.22.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.4
